added the np parameter to the dif func to don't delete th files that are not to be deleted

+++ temp_del.py
@@ -2,11 +2,16 @@
 import shutil
 import os
 import time
-def dif(folders,show=False):
-    # dif : deletes all inside the folders
+def dif(folders,show=False,na=[]):
+    """
+    deletes all the files and folders inside the folders
+    parameters:
+    show : shows the files/folders deleted
+    na: enter files not to be deleted
+    """
     tlen=0
     tdel=0
-    if show==True:
+    if show:
         deleted={'file':[],'folders':[]}
     for path in folders:
         os.chdir(path)
@@ -14,27 +19,29 @@ def dif(folders,show=False):
         print("IN :",os.getcwd())
         for x in os.listdir():
             try:
-                if os.path.isfile(x):
-                    os.remove(x)
-                    tdel+=1
-                    if show==True:
-                        deleted['file'].append(x)
+                if x not in na:
+                    if os.path.isfile(x):
+                    
+                        os.remove(x)
+                        tdel+=1
+                        if show:
+                            deleted['file'].append(x)
 
-                else:
-                    shutil.rmtree(x)
-                    tdel+=1
-                    if show==True:
-                        deleted['folders'].append(x)
+                    else:
+                        shutil.rmtree(x)
+                        tdel+=1
+                        if show:
+                            deleted['folders'].append(x)
             except:
                 pass
     
     print('deleted: {} out of total: {} '.format(tdel,tlen))
-    if show==True:
+    if show:
          print('files deleted : {} \n folders deleted : {} '.format(tuple(deleted['file']),tuple(deleted['folders'])))
 
 def main():
         #delo(("C:/Windows/Temp","C:/Users/Admin/AppData/Local/Temp"))
-        dif(('C:/Users/Admin/OneDrive/Desktop/b','C:/Users/Admin/OneDrive/Desktop/a'),show=True)
+        dif(('C:/Users/Admin/OneDrive/Desktop/b','C:/Users/Admin/OneDrive/Desktop/a'),show=True,na=['kacha aloo'])
         time.sleep(0.5)
 
 if __name__=="__main__":
